pyproject.toml
[project]
name = "policy-agent"
version = "0.1.0"
description = "ADK agent that retrieves policy documents from GCS using Agent Identity, deployable as a Gemini Enterprise add-on agent."
requires-python = ">=3.10"
dependencies = [
  "google-cloud-aiplatform[adk,agent_engines]>=1.95.0",
  "google-adk>=1.0.0",
  "google-cloud-storage>=2.18.0",
  "google-cloud-secret-manager>=2.20.0",
  "google-auth>=2.35.0",
  "pydantic>=2.7.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "ruff>=0.6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["policy_agent"]
